requirements.txt
boto3>=1.28.57
awscli>=1.29.57
botocore>=1.31.57
qdrant-client
sentence-transformers
streamlit
python-dotenv
opensearch-py
pandas
requests-aws4auth
PyPDF2